% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing.R
\name{print.safe_multiple_test}
\alias{print.safe_multiple_test}
\title{Print Method for Multiple Testing Results}
\usage{
\method{print}{safe_multiple_test}(x, ...)
}
\arguments{
\item{x}{A safe_multiple_test object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Multiple Testing Results
}
