% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_coef_test.R
\name{print.safe_test_result}
\alias{print.safe_test_result}
\title{Print Method for Safe Test Results}
\usage{
\method{print}{safe_test_result}(x, ...)
}
\arguments{
\item{x}{A safe_test_result object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Safe Test Results
}
